Remove try_from f64, Add try_from &str, Fix try_from (f64, rounding)
Constructing an asset from float is not safe.
For example if we have an asset called USD with a precision of 4 digit after the decimal,
constructing a new Credit from 10.87, seems a safe operation but is not, becasue the value
of 10.87 is parsed as:

  exponent b10000000010 -> 1026 -> 1206 - 1023 -> 3
  mantissa b0101101111010111000010100011110101110000101000111101 ->
           b1.0101101111010111000010100011110101110000101000111101
  mantissa * 2^exp = b1010.1101111010111000010100011110101110000101000111101 ->
                     10.8699999999999992184

So a number that seems to have a precision of 2 digit after the decimal is parsed in a
number with a precision of 19 digit after the decimal.

For that:
1. Contructing an asset from a plain floating point number is now no more possible
because either Merx should fail for all too precise value or Merx should round too precises
value in an arbitrary way. Both the options are not acceptable.
2. When the source of the float is a text string the best thing to do is to parse the value
directly from the text string.